# docs/openapi.yaml
openapi: 3.1.0
info:
  title: RinoTravel API
  version: 1.0.0
  description: |
    API do RinoTravel. Fases implementadas: fundação, autenticação, viagens com membros, roteiro, lugares e restaurantes, voos e hotéis, transfers, documentos, sync offline-first e adaptadores Google.

    Erros usam `application/problem+json` (RFC 9457). O campo `code` é estável e pensado para o cliente traduzir a mensagem.
    Quem não é membro de uma viagem recebe sempre `404 trip_not_found`, idêntico ao de uma viagem inexistente.
servers:
  - url: http://localhost:8080
    description: Desenvolvimento local
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/trips/{tripId}/places:
    parameters:
      - $ref: '#/components/parameters/TripId'
    get:
      tags: [Places]
      operationId: listPlaces
      summary: Lista os lugares (lista de desejos)
      responses:
        '200':
          description: Lista de os lugares (lista de desejos).
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Place'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Places]
      operationId: createPlace
      summary: Cria os lugares (lista de desejos) (OWNER, ADMIN, MEMBER). O id pode ser gerado pelo cliente (UUID v7), o que torna a criação idempotente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaceCreate'
      responses:
        '201':
          description: Criado.
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Place'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
...
  /api/v1/trips/{tripId}/restaurants:
    parameters:
      - $ref: '#/components/parameters/TripId'
    get:
      tags: [Places]
      operationId: listRestaurants
      summary: Lista os restaurantes
      responses:
        '200':
          description: Lista de os restaurantes.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Restaurant'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Places]
      operationId: createRestaurant
      summary: Cria os restaurante (OWNER, ADMIN, MEMBER). O id pode ser gerado pelo cliente (UUID v7), o que torna a criação idempotente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestaurantCreate'
      responses:
        '201':
          description: Criado.
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
...
  /api/v1/places/search:
    get:
      tags: [Places]
      operationId: searchPlaces
      summary: Busca lugares no Google Places. Só existe quando o servidor tem GOOGLE_MAPS_API_KEY
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: lat
          in: query
          schema:
            type: number
        - name: lng
          in: query
          schema:
            type: number
        - name: language
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Candidatos.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/PlaceCandidate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/Unavailable'
...
  /api/v1/trips/{tripId}/maps/location:
    parameters:
      - $ref: '#/components/parameters/TripId'
    post:
      tags: [Places]
      operationId: renderLocationMap
      summary: Desenha o mapa de um único local (item do roteiro, lugar, restaurante ou hotel). Só existe quando o servidor tem GOOGLE_MAPS_API_KEY
      description: |
        Devolve uma imagem PNG com o local marcado. A imagem é feita na hora e nada dela é guardada no servidor.
        Qualquer membro da viagem (inclusive LEITOR) pode pedir.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationMapRequest'
      responses:
        '200':
          description: A imagem do mapa (cache privado de 5 minutos).
          content:
            image/png:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/Unavailable'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Token de sessão devolvido por register e login (prefixo `rt_`, validade de 30 dias).
  parameters:
    TripId:
      name: tripId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    BadRequest:
      description: Requisição malformada (JSON inválido, campo desconhecido, id fora do formato).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Sem autenticação, token inválido ou expirado, ou credenciais inválidas.
      headers:
        WWW-Authenticate:
          schema:
            type: string
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Autenticado, mas sem permissão para esta ação.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Recurso inexistente, ou que o usuário não pode saber que existe.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: Conflito de estado (versão desatualizada, email já cadastrado, id já usado, membro já existe).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unprocessable:
      description: Dados bem formados que violam uma regra; `errors` lista os campos.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    TooManyRequests:
      description: Limite de requisições excedido.
      headers:
        Retry-After:
          description: Segundos até poder tentar de novo.
          schema:
            type: integer
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unavailable:
      description: O provedor externo (Google) está indisponível, não respondeu a tempo, ou o limite mensal de chamadas foi atingido (`provider_quota_exhausted`).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    ErrorCode:
      type: string
      description: Código estável do erro; use-o para traduzir a mensagem.
      enum:
        - route_not_found
        - malformed_json
        - body_too_large
        - invalid_id
        - unauthenticated
        - invalid_token
        - invalid_credentials
        - invalid_registration_code
        - email_taken
        - validation_failed
        - rate_limited
        - internal_error
        - user_not_found
        - trip_not_found
        - trip_id_taken
        - version_conflict
        - empty_patch
        - forbidden
        - member_exists
        - member_not_found
        - owner_locked
        - owner_assignment
        - already_owner
        - base_version_required
        - day_exists
        - day_not_empty
        - document_not_found
        - entity_deleted
        - invalid_cursor
        - invalid_lat
        - invalid_lng
        - invalid_limit
        - invalid_link
        - invalid_mutation
        - invalid_payload
        - item_not_found
        - link_expired
        - mutation_id_reused
        - object_not_found
        - place_not_found
        - provider_unavailable
        - provider_quota_exhausted
        - size_mismatch
        - too_many_mutations
        - unknown_entity
        - unsupported_operation
        - upload_mismatch
        - upload_missing
    Problem:
      type: object
      required: [status, title, code, detail]
      properties:
        status:
          type: integer
        title:
          type: string
        code:
          $ref: '#/components/schemas/ErrorCode'
        detail:
          type: string
          description: Mensagem em inglês, segura para exibir como fallback.
        requestId:
          type: string
          description: Identificador para suporte; o mesmo vai no header X-Request-ID.
        errors:
          type: array
          description: Presente em erros de validação (422).
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
              message:
                type: string
...
    ZonedTime:
      type: object
      description: Data e hora locais no fuso informado. O instante absoluto é derivado dos dois campos.
      additionalProperties: false
      required: [dateTime, timezone]
      properties:
        dateTime:
          type: string
          example: '2027-04-10T09:30:00'
        timezone:
          type: string
          description: Nome IANA do fuso.
          example: Asia/Tokyo
    Location:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
    Money:
      type: object
      description: Valor em unidades menores da moeda (centavos) e código ISO 4217.
      required: [amount, currency]
      properties:
        amount:
          type: integer
          format: int64
        currency:
          type: string
          example: JPY
    ResourceMeta:
      type: object
      required: [id, tripId, version, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        tripId:
          type: string
          format: uuid
        version:
          type: integer
          format: int64
          description: Enviada de volta como `baseVersion` para edições.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PlaceCategory:
      type: string
      description: Categoria de um lugar.
      enum: [ATTRACTION, RESTAURANT, SHOPPING, OTHER]
    Priority:
      type: string
      description: Prioridade na lista de desejos.
      enum: [HIGH, MEDIUM, LOW]
    RestaurantStatus:
      type: string
      description: Estado de um restaurante.
      enum: [WISHLIST, PLANNED, RESERVED, VISITED]
...
    Place:
      allOf:
        - $ref: '#/components/schemas/ResourceMeta'
        - type: object
          required: [name, category, priority]
          properties:
            name:
              type: string
            description:
              type: string
            category:
              $ref: '#/components/schemas/PlaceCategory'
            priority:
              $ref: '#/components/schemas/Priority'
            location:
              $ref: '#/components/schemas/Location'
            estimatedDurationMinutes:
              type: integer
            estimatedCost:
              $ref: '#/components/schemas/Money'
            notes:
              type: string
            externalId:
              type: string
              description: Id do lugar no provedor (Google Places), quando veio de uma busca.
    PlaceCreate:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        id:
          type: string
          format: uuid
          description: Opcional. Gerado pelo cliente (UUID v7) para criação idempotente.
        name:
          type: string
        description:
          type: string
        category:
          $ref: '#/components/schemas/PlaceCategory'
        priority:
          $ref: '#/components/schemas/Priority'
        notes:
          type: string
        externalId:
          type: string
        location:
          oneOf:
            - $ref: '#/components/schemas/Location'
            - type: 'null'
        estimatedDurationMinutes:
          oneOf:
            - type: integer
            - type: 'null'
        estimatedCost:
          oneOf:
            - $ref: '#/components/schemas/Money'
            - type: 'null'
...
    Restaurant:
      allOf:
        - $ref: '#/components/schemas/ResourceMeta'
        - type: object
          required: [name, status, desiredDishes]
          properties:
            name:
              type: string
            cuisine:
              type: string
            status:
              $ref: '#/components/schemas/RestaurantStatus'
            location:
              $ref: '#/components/schemas/Location'
            estimatedCost:
              $ref: '#/components/schemas/Money'
            reservationAt:
              $ref: '#/components/schemas/ZonedTime'
            reservationCode:
              type: string
              description: Omitido para VIEWER.
            desiredDishes:
              type: array
              items:
                type: string
            notes:
              type: string
            externalId:
              type: string
    RestaurantCreate:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        id:
          type: string
          format: uuid
          description: Opcional. Gerado pelo cliente (UUID v7) para criação idempotente.
        name:
          type: string
        cuisine:
          type: string
        notes:
          type: string
        status:
          $ref: '#/components/schemas/RestaurantStatus'
        externalId:
          type: string
        reservationCode:
          type: string
        desiredDishes:
          type: array
          items:
            type: string
        location:
          oneOf:
            - $ref: '#/components/schemas/Location'
            - type: 'null'
        estimatedCost:
          oneOf:
            - $ref: '#/components/schemas/Money'
            - type: 'null'
        reservationAt:
          oneOf:
            - $ref: '#/components/schemas/ZonedTime'
            - type: 'null'
...
    PlaceCandidate:
      type: object
      required: [providerId, name, types]
      properties:
        providerId:
          type: string
        name:
          type: string
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        types:
          type: array
          items:
            type: string
...
    LocationMapRequest:
      type: object
      additionalProperties: false
      required: [location]
      properties:
        location:
          $ref: '#/components/schemas/Location'
        language:
          type: string
...